pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "hubstudio-sign-outlook"
version = "0.1.0"
description = "Hubstudio CDP + Playwright: Outlook signup page Phase 1"
requires-python = ">=3.10"
dependencies = [
    "playwright>=1.40",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-ra -q"
